Move KNNCodecService to codec folder
